Make clicking “Leave Feedback” draft email to support@rainbow.me
