+++ src/hoc/index.js
@@ -18,6 +18,7 @@ export { default as withRequests } from './withRequests';
 export { default as withRequestsInit } from './withRequestsInit';
 export { default as withRotationForDirection } from './withRotationForDirection';
 export { default as withSafeAreaViewInsetValues } from './withSafeAreaViewInsetValues';
+export { default as withSendFeedback } from './withSendFeedback';
 export { default as withTransactionConfirmationScreen } from './withTransactionConfirmationScreen';
 export { default as withTransitionProps } from './withTransitionProps';
 export { default as withWalletConnectConnections } from './withWalletConnectConnections';
